Add Cloudflare Modules E2E test
