+++ e2e/tests/cf-modules.ts
@@ -1,0 +1,3 @@
+import { createCFDeployment } from './create-cf-deployment'
+
+export const cfModulesDeployment = createCFDeployment('cloudflare-modules')
